pyproject.toml
[tool.poetry]
name = "omie-interface"
version = "0.1.0"
description = ""
authors = ["felipe <felipelhora@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
flask = "^3.0.3"
python-decouple = "^3.8"
flask-cors = "^5.0.0"
gunicorn = "^23.0.0"
requests = "^2.32.3"
uvicorn = "^0.31.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
